requirements.txt
streamlit
google-generativeai
pandas
numpy
scikit-learn
nltk
Pillow
